requirements.txt
absl-py==2.3.1
amqp==5.3.1
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.11.0
astroid==4.0.2
asttokens==3.0.1
astunparse==1.6.3
async-lru==2.0.5
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
autopep8==2.0.4
babel==2.17.0
beautifulsoup4==4.14.3
billiard==4.2.4
black==25.12.0
bleach==6.3.0
cbor2==5.7.1
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
comm==0.2.3
constantly==23.10.4
contourpy==1.3.3
crispy-bootstrap4==2024.1
cron_descriptor==2.0.6
cryptography==46.0.3
cycler==0.12.1
daphne==4.2.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
dj-database-url==2.1.0
Django==5.0.4
django-allauth==65.13.1
django-appconf==1.2.0
django-celery-beat==2.8.1
django-channels==0.7.0
django-ckeditor==6.7.1
django-cloudinary-storage==0.3.0
django-crispy-forms==2.1
django-filter==24.2
django-jazzmin==3.0.0
django-js-asset==3.1.2
django-phone-field==1.8.1
django-phonenumber-field==7.3.0
django-redis==6.0.0
django-robots==6.1
django-smart-selects==1.7.2
django-timezone-field==7.2.1
django_celery_results==2.6.0
django_compressor==4.6.0
docstring-to-markdown==0.17
environs==14.1.1
exceptiongroup==1.3.1
executing==2.2.1
Faker==39.0.0
fastjsonschema==2.21.2
fuzzywuzzy==0.18.0
Levenshtein==0.26.1
filelock==3.19.1
flake8==7.1.2
flatbuffers==25.9.23
fonttools==4.61.0
fqdn==1.5.1
fsspec==2025.9.0
gast==0.7.0
google-pasta==0.2.0
grpcio==1.76.0
gunicorn==21.2.0


MarkupSafe==3.0.3
marshmallow==4.0.0
matplotlib==3.10.7
matplotlib-inline==0.2.1
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.4
ml_dtypes==0.5.4
mpmath==1.3.0
msgpack==1.1.2
mypy_extensions==1.1.0
namex==0.1.0
nbclassic==1.3.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook_shim==0.2.4
numpy==2.2.6
oauthlib==3.3.1
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
phonenumbers==8.13.40
pillow==12.0.0

protobuf==6.33.2
psutil==7.1.3
psycopg2-binary==2.9.11

pyflakes==3.2.0
Pygments==2.19.2
PyJWT==2.8.0
pylint==4.0.4

PyYAML==6.0.3
pyzmq==27.1.0
rcssmin==1.2.2
redis==7.1.0
referencing==0.37.0
requests==2.31.0
requests-oauthlib==2.0.0
resend==2.19.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rjsmin==1.2.5
rope==1.14.0
rpds-py==0.30.0
scikit-learn==1.7.2
scipy==1.16.3
Send2Trash==1.8.3
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
snowballstemmer==3.0.1
soupsieve==2.8
sqlparse==0.5.4
stack-data==0.6.3

Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.11.0
uri-template==1.3.0
urllib3==2.6.0
vine==5.1.0
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.4
whatthepatch==1.0.7
wheel==0.45.1
whitenoise==6.11.0
widgetsnbextension==4.0.15
wrapt==2.0.1
yapf==0.43.0
zipp==3.23.0
zope.interface==8.1.1
django-embed-video==1.4.10
